backend/requirements.txt
livekit-agents[google]
livekit-plugins-silero
python-dotenv
livekit-api
flask[async]
flask
flask-cors
uvicorn
psycopg2-binary
supabase

# RAG Dependencies for Resume Analysis
langchain
langchain-community
langchain-google-genai
langchain-chroma
chromadb
sentence-transformers
pypdf2
python-docx
markdown

# Resume Processing
pdfplumber
pyresparser
spacy
nltk

# File Upload
werkzeug